Created Level 1 and added comments
I created a basic first level but there is still no way to win it, nor is there a use for Shrek.

I also added comments to the code where I could. Some of the collision logic I do not under stand and some one will have explain that for me.